Revert accidental package renaming
