+++ core/src/main/java/com/example/demo/core/InvalidStockQuantity.java
@@ -1,4 +1,4 @@
-package com.example.demo.demo;
+package com.example.demo.core;
 
 public class InvalidStockQuantity extends RuntimeException {
