+++ core/src/main/java/com/example/demo/core/ShoeCore.java
@@ -1,4 +1,4 @@
-package com.example.demo.demo;
+package com.example.demo.core;
 
 import com.example.demo.dto.in.ShoeFilter;
 import com.example.demo.dto.out.Shoes;
