+++ core/src/main/java/com/example/demo/core/StockEntity.java
@@ -1,4 +1,4 @@
-package com.example.demo.demo;
+package com.example.demo.core;
 
 public record StockEntity(int quantity) {
