+++ core/src/main/java/com/example/demo/core/StockService.java
@@ -1,4 +1,4 @@
-package com.example.demo.demo;
+package com.example.demo.core;
 
 import com.example.demo.dto.out.Stock;
 import com.example.demo.dto.out.Stock.StockState;
