+++ core/src/main/java/com/example/demo/facade/ShoeFacade.java
@@ -1,6 +1,6 @@
 package com.example.demo.facade;
 
-import com.example.demo.demo.ShoeCore;
+import com.example.demo.core.ShoeCore;
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.stereotype.Component;
